requirements.txt
googlemaps
telethon
python-dotenv
ollama